fix: use alembic stamp instead of upgrade for staging
staging database tables are created by init_database() in the fastapi lifespan,
so migrations that try to ALTER tables will fail if the base tables don't exist.

using stamp instead of upgrade marks the database as being at the latest migration
without actually running the migrations, which is correct since the schema matches.